refactor(afericoes): pagina busca das aferições na integração com Avilan

+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/AfericaoDao.java
@@ -60,10 +60,10 @@ public interface AfericaoDao {
      * @return uma lista de aferições
      * @throws SQLException se ocorrer erro no banco
      */
-    List<Afericao> getAfericoesByCodUnidadeByPlaca(List<String> codUnidades, List<String> placas, long limit, long offset) throws SQLException;
+    List<Afericao> getAfericoesByCodUnidadeByPlaca(List<String> codUnidades, List<String> placas, int limit, long offset) throws SQLException;
 
     List<Afericao> getAfericoes(Long codUnidade, String codTipoVeiculo, String placaVeiculo, long dataInicial,
-                                long dataFinal, long limit, long offset) throws SQLException;
+                                long dataFinal, int limit, long offset) throws SQLException;
 
     /**
      * retorna uma aferição através do código dela
